+++ scripts/add_brownfield.py
@@ -11,7 +11,7 @@ import yaml
 import numpy as np
 
 from add_existing_baseyear import add_build_year_to_new_assets
-from helper import override_component_attrs
+from helper import override_component_attrs, update_config_with_sector_opts
 from solve_network import basename
 
 
@@ -123,6 +123,8 @@ if __name__ == "__main__":
             planning_horizons=2030,
         )
 
+    update_config_with_sector_opts(snakemake.config, snakemake.wildcards.sector_opts)
+
     print(snakemake.input.network_p)
     logging.basicConfig(level=snakemake.config['logging_level'])
 
@@ -137,4 +139,5 @@ if __name__ == "__main__":
 
     add_brownfield(n, n_p, year)
 
+    n.meta = dict(snakemake.config, **dict(wildcards=dict(snakemake.wildcards)))
     n.export_to_netcdf(snakemake.output[0])