+++ scripts/helper.py
@@ -1,7 +1,9 @@
 import os
+import yaml
 import pytz
 import pandas as pd
 from pathlib import Path
+from snakemake.utils import update_config
 from pypsa.descriptors import Dict
 from pypsa.components import components, component_attrs
 
@@ -23,7 +25,7 @@ def override_component_attrs(directory):
 
     Returns
     -------
-    Dictionary of overriden component attributes.
+    Dictionary of overridden component attributes.
     """
 
     attrs = Dict({k : v.copy() for k,v in component_attrs.items()})
@@ -123,3 +125,17 @@ def generate_periodic_profiles(dt_index, nodes, weekly_profile, localize=None):
     week_df = week_df.tz_localize(localize)
 
     return week_df
+
+
+def parse(l):
+    if len(l) == 1:
+        return yaml.safe_load(l[0])
+    else:
+        return {l.pop(0): parse(l)}
+
+
+def update_config_with_sector_opts(config, sector_opts):
+    for o in sector_opts.split("-"):
+        if o.startswith("CF:"):
+            l = o.split("+")[1:]
+            update_config(config, parse(l))